+++ ntex-net/src/compio/io_impl.rs
@@ -36,6 +36,8 @@ impl Handle for HandleWrapper {
         }
         None
     }
+
+    fn notify(&self, _: &IoContext) {}
 }
 
 #[allow(dead_code)]